refactored directories and names
